fix: Resolve deployment issues
This commit resolves the following deployment issues:
- Moves the swipe page to the correct location (`/swipe`)
- Adds a redirect from the root to the swipe page
- Fixes the database connection issue by correctly accessing the `MONGODB_URI` environment variable
- Fixes the build error caused by incorrect import paths
