--- src/lib/config.ts
@@ -1,3 +1,0 @@
-import "dotenv/config";
-
-export const MONGODB_URI = process.env.MONGODB_URI || "";
